+++ docs/parsers/history.md
@@ -1,4 +1,9 @@
 [Home](https://kellyjonbrazil.github.io/jc/)
+# Table of Contents
+
+* [jc.parsers.history](#jc.parsers.history)
+  * [parse](#jc.parsers.history.parse)
+
 <a id="jc.parsers.history"></a>
 
 # jc.parsers.history